CMake config file
A CMake library project should include a cmake config file for easier use from
other cmake projects. See:
https://cmake.org/Wiki/CMake/Tutorials/How_to_create_a_ProjectConfig.cmake_file

In contrast to the FindXXX.cmake script that wildly guesses where the
required files of the g2o are, if g2o is <make install>'d cmake knows the
location of the libraries and headers and can make that knowledge available
with the config file.

The user can then install g2o (and other cmake projects) to an arbitrary
location, e.g. ~/local or ~/slam_projects, and tell cmake to check there for
the config files by setting the CMAKE_PREFIX_PATH.